@@ -1,26 +1,20 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using Microsoft.AspNetCore.Builder;
+using Rcm.Services.Users.Api;
+using Rcm.Services.Users.Core;
+using Rcm.Shared.Middlewares;
 
-namespace Rcm.Services.Users.Api
+internal class Program
 {
-    public class Program
+    private static void Main(string[] args)
     {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+        var builder = WebApplication.CreateBuilder(args);
+        builder.Services.AddCore();
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
+        var app = builder.Build();
+        app.UseMiddleware<ErrorHandlerMiddleware>();
+
+        app.MapEndpoints();
+
+        app.Run();
     }
 }
