@@ -3,14 +3,7 @@ using System.Threading.Tasks;
 using MediatR;
 using Rcm.Services.Users.Core.DAL;
 
-namespace Rcm.Services.Users.Core.Features.Accounts.Commands.SignUp;
-
-public class SignUpCommand : IRequest<int>
-{
-    public string Email { get; set; }
-    public string Password { get; set; }
-}
-
+namespace Rcm.Services.Users.Core.Commands.SignUp;
 public class SignUpCommandHandler : IRequestHandler<SignUpCommand, int>
 {
     private readonly UsersDbContext _context;
